@@ -16,62 +16,58 @@
   "devDependencies": {
     "@types/highcharts": "^4.2.46",
     "@types/path-is-absolute": "^1.0.0",
-    "@types/plotly.js": "0.0.5",
     "@types/pluralize": "^0.0.27",
-    "@types/react": "^15.0.31",
-    "@types/react-bootstrap": "0.0.51",
-    "@types/react-dom": "^15.5.1",
-    "@types/react-fontawesome": "^1.5.1",
-    "@types/react-responsive": "^1.1.7",
-    "@types/react-router": "^3.0.0",
-    "@types/react-select": "^1.0.48",
-    "@types/react-sidebar": "^2.2.0-next.0",
+    "@types/react": "^16.0.22",
+    "@types/react-bootstrap": "^0.31.7",
+    "@types/react-dom": "^16.0.3",
+    "@types/react-responsive": "^3.0.0",
+    "@types/react-router-dom": "^4.2.1",
+    "@types/react-select": "^1.0.59",
     "@types/react-syntax-highlighter": "^0.0.3",
     "@types/react-table": "^6.6.0",
     "@types/webpack": "^2.2.6",
     "@types/webpack-dev-server": "^2.4.0",
     "file-loader": "^1.1.5",
     "gulp": "^3.9.1",
-    "gulp-shell": "^0.6.3"
+    "gulp-shell": "^0.6.3",
+    "url-loader": "^0.6.2"
   },
   "dependencies": {
-    "apollo-client": "^1.3.0",
+    "apollo-cache-inmemory": "^1.1.0",
+    "apollo-client": "^2.0.2",
+    "apollo-link-http": "^1.1.0",
     "css-loader": "^0.28.4",
     "cytoscape": "^2.7.13",
-    "deep-equal": "^1.0.1",
-    "graphql": "^0.10.3",
-    "graphql-tag": "^2.4.2",
+    "graphql": "^0.11.7",
+    "graphql-tag": "^2.5.0",
     "highcharts": "^5.0.7",
     "highcharts-heatmap": "^0.1.2",
     "highcharts-more": "^0.1.2",
     "moment": "^2.17.1",
     "ndb-react-components": "^0.1.27",
     "numeric": "^1.2.6",
-    "plotly.js": "^1.21.3",
     "pluralize": "^6.0.0",
     "rc-slider": "^8.2.0",
-    "react": "^15.6.1",
-    "react-addons-shallow-compare": "^15.4.2",
-    "react-apollo": "1.2.0",
+    "react": "^16.0.0",
+    "react-apollo": "^2.0.1",
     "react-autosuggest": "^9.3.1",
-    "react-bootstrap": "^0.31.0",
-    "react-datepicker": "^0.48.0",
-    "react-dom": "^15.6.1",
-    "react-fontawesome": "^1.6.1",
+    "react-bootstrap": "^0.31.5",
+    "react-datepicker": "^0.60.2",
+    "react-dom": "^16.0.0",
     "react-highcharts": "^11.5.0",
-    "react-responsive": "^1.3.0",
-    "react-router": "^3.0.2",
-    "react-select": "^1.0.0-rc.5",
-    "react-sidebar": "^2.2.1",
+    "react-responsive": "^3.0.0",
+    "react-router-dom": "^4.2.2",
+    "react-select": "^1.0.0-rc.10",
     "react-syntax-highlighter": "^5.6.2",
     "react-table": "^6.7.4",
-    "react-tap-event-plugin": "^1.0.0",
-    "react-toastify": "^1.7.0",
-    "react-virtualized": "^9.8.0",
-    "react-virtualized-select": "^3.0.1",
+    "react-toastify": "^2.2.0",
+    "react-virtualized": "^9.12.0",
+    "react-virtualized-select": "^3.1.0",
+    "semantic-ui": "^2.2.13",
+    "semantic-ui-react": "^0.76.0",
     "style-loader": "^0.18.2",
     "ts-loader": "^2.2.0",
-    "typescript": "^2.4.0",
+    "typescript": "2.4.2",
     "webpack": "^2.6.1",
     "webpack-dev-server": "^2.5.0"
   }
