@@ -1,23 +1,20 @@
 import * as React from "react";
-
-import {Loading} from "../../Loading";
+import {Loader} from "semantic-ui-react";
 import {graphql} from "react-apollo";
-import {contentStyles} from "../../util/styleDefinitions";
+
 import {ProjectsQuery} from "../../graphql/project";
 import {ProjectsPanel} from "./ProjectsPanel";
 
-
-@graphql(ProjectsQuery, {
-    options: {
-        pollInterval: 5 * 1000
-    }
-})
-export class Projects extends React.Component<any, any> {
+export class _Projects extends React.Component<any, any> {
     public render() {
         const loading = !this.props.data || this.props.data.loading;
 
         if (loading) {
-            return (<Loading/>);
+            return (
+                <div style={{display: "flex", height: "100%", alignItems: "center"}}>
+                    <Loader active inline="centered">Loading</Loader>
+                </div>
+            );
         }
 
         if (this.props.data.error) {
@@ -25,9 +22,13 @@ export class Projects extends React.Component<any, any> {
         }
 
         return (
-            <div style={contentStyles.body}>
-                <ProjectsPanel projects={this.props.data.projects}/>
-            </div>
+            <ProjectsPanel projects={this.props.data.projects}/>
         );
     }
 }
+
+export const Projects = graphql<any, any>(ProjectsQuery, {
+    options: {
+        pollInterval: 5 * 1000
+    }
+})(_Projects);
