@@ -1,15 +1,12 @@
 import * as React from "react";
 import {graphql} from "react-apollo";
+import {Loader} from "semantic-ui-react";
 
-import {Loading} from "../../Loading";
-import {contentStyles} from "../../util/styleDefinitions";
 import {PipelineStagesPanel} from "./PipelineStagesPanel";
 import {PipelineStagesQuery} from "../../graphql/pipelineStage";
 import {ProjectsQuery} from "../../graphql/project";
 import {IPipelineStage} from "../../models/pipelineStage";
 import {PipelineStageDetails} from "./details/PipelineStageDetails";
-
-const styles = contentStyles;
 
 interface IPipelineStagesProps {
     projectsData?: any;
@@ -21,18 +18,7 @@ interface IPipelineStagesState {
     selectedStage?: IPipelineStage;
 }
 
-@graphql(ProjectsQuery, {
-    name: "projectsData",
-    options: {
-        pollInterval: 5 * 1000
-    }
-})
-@graphql(PipelineStagesQuery, {
-    options: {
-        pollInterval: 5 * 1000
-    }
-})
-export class PipelineStages extends React.Component<IPipelineStagesProps, IPipelineStagesState> {
+class __PipelineStages extends React.Component<IPipelineStagesProps, IPipelineStagesState> {
     constructor(props) {
         super(props);
         this.state = {pipelinesForProjectId: "", selectedStage: null};
@@ -50,19 +36,23 @@ export class PipelineStages extends React.Component<IPipelineStagesProps, IPipel
         const loading = !this.props.data || this.props.data.loading || !this.props.projectsData || this.props.projectsData.loading;
 
         if (loading) {
-            return (<Loading/>);
+            return (
+                <div style={{display: "flex", height: "100%", alignItems: "center"}}>
+                    <Loader active inline="centered">Loading</Loader>
+                </div>
+            );
         }
 
         if (this.props.data.error) {
-            return (<span>{this.props.data.error}</span>);
+            return this.props.data.error;
         }
 
         if (this.props.projectsData.error) {
-            return (<span>{this.props.projectsData.error}</span>);
+            return this.props.projectsData.error;
         }
 
         return (
-            <div style={styles.body}>
+            <div>
                 <PipelineStagesPanel projects={this.props.projectsData.projects}
                                      pipelineStages={this.props.data.pipelineStages}
                                      pipelinesForProjectId={this.state.pipelinesForProjectId}
@@ -74,3 +64,16 @@ export class PipelineStages extends React.Component<IPipelineStagesProps, IPipel
         );
     }
 }
+
+const _PipelineStages = graphql<any, any>(ProjectsQuery, {
+    name: "projectsData",
+    options: {
+        pollInterval: 5 * 1000
+    }
+})(__PipelineStages);
+
+export const PipelineStages = graphql<any, any>(PipelineStagesQuery, {
+    options: {
+        pollInterval: 5 * 1000
+    }
+})(_PipelineStages);
