@@ -1,25 +1,17 @@
 import * as React from "react";
-import {Pagination, ControlLabel, Row, Col} from "react-bootstrap"
+import {Container, Menu} from "semantic-ui-react";
 
 import {graphql} from "react-apollo";
 import gql from "graphql-tag";
 
-import {ITaskExecution} from "../../../models/taskExecution";
 import {IPipelineStage} from "../../../models/pipelineStage";
 import {TilesTable} from "./TilesTable";
 import {
-    TILE_PIPELINE_STATUS_FAILED, TILE_PIPELINE_STATUS_TYPES,
+    TILE_PIPELINE_STATUS_TYPES,
     TilePipelineStatusType
 } from "../../../models/tilePipelineStatus";
 import {TilePipelineStatusSelect} from "../../helpers/TilePipelineStatusSelect";
-
-export interface ITilesPage {
-    offset: number;
-    limit: number;
-    totalCount: number;
-    hasNextPage: boolean;
-    items: ITaskExecution[]
-}
+import {PreferencesManager} from "../../../util/preferencesManager";
 
 interface ITilesProps {
     pipelineStage: IPipelineStage;
@@ -28,8 +20,6 @@ interface ITilesProps {
 interface ITilesState {
     offset?: number;
     limit?: number;
-    totalCount?: number;
-    requestedOffset?: number;
     requestedStatus?: TilePipelineStatusType;
 }
 
@@ -40,45 +30,35 @@ export class Tiles extends React.Component<ITilesProps, ITilesState> {
         this.state = {
             offset: 0,
             limit: 20,
-            totalCount: -1,
-            requestedOffset: 0,
-            requestedStatus: TILE_PIPELINE_STATUS_FAILED
+            requestedStatus: TilePipelineStatusType.fromStatus(PreferencesManager.Instance.TilePipelineStatus)
         }
     }
 
     private onTilePipelineStatusTypeChanged(t: TilePipelineStatusType) {
+        PreferencesManager.Instance.TilePipelineStatus = t.option;
+
         this.setState({requestedStatus: t});
     }
 
-    private updateOffset(offset: number) {
-        if (offset != this.state.requestedOffset) {
-            this.setState({requestedOffset: offset});
-        }
-    }
-
-    private updateCursor(page: ITilesPage) {
-        if (page.offset !== this.state.offset || page.limit !== this.state.limit || page.totalCount !== this.state.totalCount) {
+    private updateCursor(page: number, pageSize: number) {
+        const offset = page * pageSize;
+        if (offset !== this.state.offset || pageSize !== this.state.limit) {
             this.setState({
-                offset: page.offset,
-                limit: page.limit,
-                totalCount: page.totalCount
+                offset,
+                limit: pageSize
             });
-        }
+      }
     }
 
     public render() {
+
         return (
-            <div>
-                <TablePanelWithQuery pipelineStage={this.props.pipelineStage}
-                                     offset={this.state.offset}
-                                     limit={this.state.limit}
-                                     totalCount={this.state.totalCount}
-                                     requestedOffset={this.state.requestedOffset}
-                                     requestedStatus={this.state.requestedStatus}
-                                     onUpdateOffset={(offset: number) => this.updateOffset(offset)}
-                                     onCursorChanged={(page: ITilesPage) => this.updateCursor(page)}
-                                     onRequestedStatusChanged={(t: TilePipelineStatusType) => this.onTilePipelineStatusTypeChanged(t)}/>
-            </div>
+            <TablePanelWithQuery pipelineStage={this.props.pipelineStage}
+                                 requestedStatus={this.state.requestedStatus}
+                                 offset={this.state.offset}
+                                 limit={this.state.limit}
+                                 onCursorChanged={(page: number, pageSize: number) => this.updateCursor(page, pageSize)}
+                                 onRequestedStatusChanged={(t: TilePipelineStatusType) => this.onTilePipelineStatusTypeChanged(t)}/>
         );
     }
 }
@@ -86,88 +66,61 @@ export class Tiles extends React.Component<ITilesProps, ITilesState> {
 
 interface ITilesTablePanelProps {
     data?: any;
-
-    pipelineStage: IPipelineStage;
-
     offset: number;
     limit: number;
-    totalCount: number;
-    requestedOffset: number;
+    pipelineStage: IPipelineStage;
     requestedStatus?: TilePipelineStatusType;
 
-    onUpdateOffset(offset: number): void;
-    onCursorChanged(page: ITilesPage): void;
+    onCursorChanged(page: number, pageSize: number): void;
     onRequestedStatusChanged(t: TilePipelineStatusType): void;
 }
 
 interface ITilesTablePanelState {
 }
 
-class TablePanel extends React.Component<ITilesTablePanelProps, ITilesTablePanelState> {
-    private onTilePipelineStatusTypeChanged(t: TilePipelineStatusType) {
-        this.props.onRequestedStatusChanged(t);
-    }
-
+class TilesTablePanel extends React.Component<ITilesTablePanelProps, ITilesTablePanelState> {
     public render() {
         let tilesForStage = [];
+        let pageCount = -1;
+        let page = 0;
+        let loading = true;
 
         if (this.props.data && this.props.data.tilesForStage) {
             tilesForStage = this.props.data.tilesForStage.items;
+            loading = this.props.data.loading;
+            if (this.props.data.tilesForStage.limit > 0) {
+                pageCount = Math.ceil(this.props.data.tilesForStage.totalCount / this.props.data.tilesForStage.limit);
+                page = Math.ceil(this.props.offset / this.props.data.tilesForStage.limit);
+            }
         }
-
-        const pageCount = Math.ceil(this.props.totalCount / this.props.limit);
-
-        const activePage = this.props.offset ? (Math.floor(this.props.offset / this.props.limit) + 1) : 1;
 
         return (
-            <div>
-                <div style={{padding: "10px"}}>
-                    <Row>
-                        <Col xs={2}>
-                            <ControlLabel>Status</ControlLabel>
-                            <TilePipelineStatusSelect idName="filter-mode"
-                                                      options={TILE_PIPELINE_STATUS_TYPES}
-                                                      placeholder="required"
-                                                      clearable={false}
-                                                      selectedOption={this.props.requestedStatus}
-                                                      onSelect={(v: TilePipelineStatusType) => this.onTilePipelineStatusTypeChanged(v)}/>
-                        </Col>
-                    </Row>
-                </div>
-                {pageCount > 1 ?
-                    <Pagination prev next first last ellipsis boundaryLinks items={pageCount} maxButtons={10}
-                                activePage={activePage}
-                                onSelect={(page: any) => {
-                                    this.props.onUpdateOffset(this.props.limit * (page - 1))
-                                }}/> : null}
-                {tilesForStage.length === 0 ? <NoTasks/> :
-                    <TilesTable pipelineStage={this.props.pipelineStage} tiles={tilesForStage}
-                                canSubmit={this.props.requestedStatus.canSubmit}/>}
-            </div>
+            <Container fluid style={{padding: "20px"}}>
+                <Menu size="mini">
+                    <TilePipelineStatusSelect
+                        statusTypes={TILE_PIPELINE_STATUS_TYPES}
+                        selectedStatus={this.props.requestedStatus}
+                        onSelectStatus={(t) => this.props.onRequestedStatusChanged(t)}/>
+                </Menu>
+                <TilesTable pipelineStage={this.props.pipelineStage}
+                            tiles={tilesForStage}
+                            canSubmit={this.props.requestedStatus.canSubmit}
+                            loading={loading}
+                            pageCount={pageCount}
+                            onCursorChanged={this.props.onCursorChanged}
+            />
+            </Container>
         );
-    }
-
-    public componentWillReceiveProps(nextProps: any) {
-        if (nextProps.data && nextProps.data.tilesForStage) {
-            nextProps.onCursorChanged(nextProps.data.tilesForStage);
-        }
-    }
-}
-
-class NoTasks extends React.Component
-    <any, any> {
-    public render() {
-        return (<div style={{padding: "10px"}}> There are no tiles with this status.</div>);
     }
 }
 
 const TileStatusQuery = gql`query($pipelineStageId: String, $status: Int, $offset: Int, $limit: Int) {
-        tilesForStage(pipelineStageId: $pipelineStageId, status: $status, offset: $offset, limit: $limit) {
-            offset
-            limit
-            totalCount
-            hasNextPage
-            items {
+    tilesForStage(pipelineStageId: $pipelineStageId, status: $status, offset: $offset, limit: $limit) {
+        offset
+        limit
+        totalCount
+        hasNextPage
+        items {
             relative_path
             lat_x
             lat_y
@@ -176,9 +129,9 @@ const TileStatusQuery = gql`query($pipelineStageId: String, $status: Int, $offse
     }
 }`;
 
-const TablePanelWithQuery = graphql(TileStatusQuery, {
-    options: ({pipelineStage, requestedStatus, requestedOffset, limit}) => ({
+const TablePanelWithQuery = graphql<any, any>(TileStatusQuery, {
+    options: ({pipelineStage, requestedStatus, offset, limit}) => ({
         pollInterval: 10 * 1000,
-        variables: {pipelineStageId: pipelineStage.id, status: requestedStatus.option, offset: requestedOffset, limit}
+        variables: {pipelineStageId: pipelineStage.id, status: requestedStatus.option, offset, limit}
     })
-})(TablePanel);
+})(TilesTablePanel);
