@@ -1,15 +1,12 @@
 import * as React from "react";
-import {graphql, InjectedGraphQLProps} from "react-apollo";
+import {graphql} from "react-apollo";
 
 import {Loading} from "../../Loading";
 import {TaskRepositoryPanel} from "./repository/TaskRepositoryPanel";
 import {ITaskRepository} from "../../models/taskRepository";
 import {TaskDefinitionsPanel} from "./definitions/TaskDefinitionsPanel";
 import {ITaskDefinition} from "../../models/taskDefinition";
-import {contentStyles} from "../../util/styleDefinitions";
 import {TaskQuery} from "../../graphql/taskDefinition";
-
-const styles = contentStyles;
 
 interface ITaskQueryProps {
     taskDefinitions: ITaskDefinition[];
@@ -17,18 +14,14 @@ interface ITaskQueryProps {
     pipelineVolume: string;
 }
 
-interface ITaskDefinitionPanelProps extends InjectedGraphQLProps<ITaskQueryProps> {
+interface ITaskDefinitionPanelProps {
+    data?: any
 }
 
 interface ITaskDefinitionPanelState {
 }
 
-@graphql(TaskQuery, {
-    options: {
-        pollInterval: 5 * 1000
-    }
-})
-export class TasksPanel extends React.Component<ITaskDefinitionPanelProps, ITaskDefinitionPanelState> {
+class _TasksPanel extends React.Component<ITaskDefinitionPanelProps, ITaskDefinitionPanelState> {
     public render() {
         const loading = !this.props.data || this.props.data.loading;
 
@@ -41,7 +34,7 @@ export class TasksPanel extends React.Component<ITaskDefinitionPanelProps, ITask
         }
 
         return (
-            <div style={styles.body}>
+            <div>
                 <TaskRepositoryPanel taskRepositories={this.props.data.taskRepositories} pipelineVolume={this.props.data.pipelineVolume}/>
                 <TaskDefinitionsPanel taskDefinitions={this.props.data.taskDefinitions}
                                       taskRepositories={this.props.data.taskRepositories}/>
@@ -49,3 +42,9 @@ export class TasksPanel extends React.Component<ITaskDefinitionPanelProps, ITask
         );
     }
 }
+
+export const TasksPanel = graphql<ITaskDefinitionPanelProps, ITaskDefinitionPanelState>(TaskQuery, {
+    options: {
+        pollInterval: 5 * 1000
+    }
+})(_TasksPanel);
