@@ -1,3 +1,6 @@
+import {AllProjectsId} from "../components/helpers/ProjectMenu";
+import {TilePipelineStatus} from "../models/tilePipelineStatus";
+
 export interface INotificationListener {
     preferenceChanged(name: string, value: any);
 }
@@ -42,17 +45,31 @@ export class PreferencesManager {
         })
     }
 
-    public set IsProjecTableFiltered(b: boolean) {
+    public set IsSidebarCollapsed(b: boolean) {
         if (typeof(Storage) !== undefined) {
-            localStorage.setItem(prefix + "isProjectTableFiltered", b ? "true" : "false");
+            localStorage.setItem(prefix + "isSidebarCollapsed", b ? "true" : "false");
         }
     }
 
-    public get IsProjecTableFiltered() {
+    public get IsSidebarCollapsed(): boolean {
         if (typeof(Storage) !== undefined) {
-            return localStorage.getItem(prefix + "isProjectTableFiltered") === "true";
+            return localStorage.getItem(prefix + "isSidebarCollapsed") === "true";
         } else {
             return false;
+        }
+    }
+
+    public set PreferredProjectId(id: string) {
+        if (typeof(Storage) !== undefined) {
+            sessionStorage.setItem(prefix + "preferredProjectId", id);
+        }
+    }
+
+    public get PreferredProjectId(): string {
+        if (typeof(Storage) !== undefined) {
+            return sessionStorage.getItem(prefix + "preferredProjectId");
+        } else {
+            return AllProjectsId;
         }
     }
 
@@ -70,38 +87,34 @@ export class PreferencesManager {
             localStorage.setItem(prefix + "tracingSelectionHiddenOpacity", n.toFixed(2));
         }
     }
-
-    public get TracingFetchBatchSize() {
-        if (typeof(Storage) !== undefined) {
-            return parseInt(localStorage.getItem(prefix + "tracingFetchBatchSize"));
-        } else {
-            return 0.0;
-        }
-    }
-
-    public set TracingFetchBatchSize(n: number) {
-        if (typeof(Storage) !== undefined) {
-            localStorage.setItem(prefix + "tracingFetchBatchSize", n.toFixed(0));
-        }
-    }
     */
 
-    public set ProjectTableSort(obj: any) {
+    public get TilePipelineStatus(): TilePipelineStatus {
         if (typeof(Storage) !== undefined) {
-            localStorage.setItem(prefix + "projectTableSort", JSON.stringify(obj));
+            return parseInt(localStorage.getItem(prefix + "tilePipelineStatus"));
+        } else {
+            return TilePipelineStatus.Failed;
         }
     }
 
-    public get ProjectTableSort() {
+    public set TilePipelineStatus(s: TilePipelineStatus) {
         if (typeof(Storage) !== undefined) {
-            const str = localStorage.getItem(prefix + "projectTableSort");
-
-            if (str) {
-                return JSON.parse(str);
-            }
+            localStorage.setItem(prefix + "tilePipelineStatus", s.toFixed(0));
         }
+    }
 
-        return null;
+    public set IsProjectTableFiltered(b: boolean) {
+        if (typeof(Storage) !== undefined) {
+            localStorage.setItem(prefix + "isProjectTableFiltered", b ? "true" : "false");
+        }
+    }
+
+    public get IsProjectTableFiltered(): boolean {
+        if (typeof(Storage) !== undefined) {
+            return localStorage.getItem(prefix + "isProjectTableFiltered") === "true";
+        } else {
+            return false;
+        }
     }
 
     public set ProjectTableFilter(obj: any) {
@@ -122,8 +135,80 @@ export class PreferencesManager {
         return null;
     }
 
+    public set ProjectTableSort(obj: any) {
+        if (typeof(Storage) !== undefined) {
+            localStorage.setItem(prefix + "projectTableSort", JSON.stringify(obj));
+        }
+    }
+
+    public get ProjectTableSort(): any {
+        if (typeof(Storage) !== undefined) {
+            const str = localStorage.getItem(prefix + "projectTableSort");
+
+            if (str) {
+                return JSON.parse(str);
+            }
+        }
+
+        return null;
+    }
+
+    public set IsStageTableFiltered(b: boolean) {
+        if (typeof(Storage) !== undefined) {
+            localStorage.setItem(prefix + "isStageTableFiltered", b ? "true" : "false");
+        }
+    }
+
+    public get IsStageTableFiltered(): boolean {
+        if (typeof(Storage) !== undefined) {
+            return localStorage.getItem(prefix + "isStageTableFiltered") === "true";
+        } else {
+            return false;
+        }
+    }
+
+    public set StageTableFilter(obj: any) {
+        if (typeof(Storage) !== undefined) {
+            localStorage.setItem(prefix + "stageTableFilter", JSON.stringify(obj));
+        }
+    }
+
+    public get StageTableFilter() {
+        if (typeof(Storage) !== undefined) {
+            const str = localStorage.getItem(prefix + "stageTableFilter");
+
+            if (str) {
+                return JSON.parse(str);
+            }
+        }
+
+        return null;
+    }
+
+    public set StageTableSort(obj: any) {
+        if (typeof(Storage) !== undefined) {
+            localStorage.setItem(prefix + "stageTableSort", JSON.stringify(obj));
+        }
+    }
+
+    public get StageTableSort(): any {
+        if (typeof(Storage) !== undefined) {
+            const str = localStorage.getItem(prefix + "stageTableSort");
+
+            if (str) {
+                return JSON.parse(str);
+            }
+        }
+
+        return null;
+    }
+
     private validateDefaultSettings() {
         if (typeof(Storage) !== undefined) {
+            if (!localStorage.getItem(prefix + "isSidebarCollapsed")) {
+                localStorage.setItem(prefix + "isSidebarCollapsed", "false");
+            }
+
             if (!localStorage.getItem(prefix + "projectTableSort")) {
                 localStorage.setItem(prefix + "projectTableSort", JSON.stringify([{
                     id: "created_at",
@@ -131,12 +216,35 @@ export class PreferencesManager {
                 }]));
             }
 
+            if (!localStorage.getItem(prefix + "isProjectTableFiltered")) {
+                localStorage.setItem(prefix + "isProjectTableFiltered", "false");
+            }
+
             if (!localStorage.getItem(prefix + "projectTableFilter")) {
                 localStorage.setItem(prefix + "projectTableFilter", JSON.stringify([]));
             }
 
-            if (!localStorage.getItem(prefix + "isProjectTableFiltered")) {
-                localStorage.setItem(prefix + "isProjectTableFiltered", "false");
+            if (!localStorage.getItem(prefix + "stageTableSort")) {
+                localStorage.setItem(prefix + "stageTableSort", JSON.stringify([{
+                    id: "created_at",
+                    desc: true
+                }]));
+            }
+
+            if (!localStorage.getItem(prefix + "isStageTableFiltered")) {
+                localStorage.setItem(prefix + "isStageTableFiltered", "false");
+            }
+
+            if (!localStorage.getItem(prefix + "stageTableFilter")) {
+                localStorage.setItem(prefix + "stageTableFilter", JSON.stringify([]));
+            }
+
+            if (!sessionStorage.getItem(prefix + "preferredProjectId")) {
+                sessionStorage.setItem(prefix + "preferredProjectId", AllProjectsId);
+            }
+
+            if (!localStorage.getItem(prefix + "tilePipelineStatus")) {
+                localStorage.setItem(prefix + "tilePipelineStatus", TilePipelineStatus.Failed.toFixed(0));
             }
         }
     }
