@@ -12,9 +12,9 @@ function loadServerOptions() {
     const options = configurations.production;
 
     options.host = process.env.PIPELINE_API_CLIENT_HOST || options.host;
-    options.port = process.env.PIPELINE_API_CLIENT_PORT || options.port;
+    options.port = parseInt(process.env.PIPELINE_API_CLIENT_PORT) || options.port;
     options.graphQLHostname = process.env.PIPELINE_API_HOST || options.graphQLHostname;
-    options.graphQLPort = process.env.PIPELINE_API_PORT || options.graphQLPort;
+    options.graphQLPort = parseInt(process.env.PIPELINE_API_PORT) || options.graphQLPort;
 
     return options;
 }
